fix(ui): extract history-utils to fix react-refresh lint rule
Move TimeRange type and rangeToFrom function to a separate
history-utils.ts file to comply with react-refresh/only-export-components.
